Move common samples files to library/
Preparing to reuse these files in a new sample.
