+++ samples/audio-video-player/audio_video_player.html
@@ -54,7 +54,7 @@
   <canvas width=1280 height=720></canvas>
 </div>
 <script type="module">
-import { WebAudioController } from "./web_audio_controller.js";
+import { WebAudioController } from "../library/web_audio_controller.js";
 
 // Transfer canvas to offscreen. Painting will be performed by worker without
 // blocking the Window main thread.
