+++ samples/index.html
@@ -15,8 +15,8 @@
         <h1><a href="./video-decode-display/">Video Decoding and Display</a></h1>
         <p>Demuxes (using mp4box.js) and decodes an mp4 file, paints to canvas ASAP</p>
       </article>
-      <article onclick="window.location.href='https://webcodecs-samples.netlify.app/audio-video-player/audio_video_player.html';">
-        <h1><a href="https://webcodecs-samples.netlify.app/audio-video-player/audio_video_player.html">Audio And Video Player</a></h1>
+      <article onclick="window.location.href='./audio-video-player/audio_video_player.html';">
+        <h1><a href="./audio-video-player/audio_video_player.html">Audio And Video Player</a></h1>
         <p>Demuxes, decodes, and renders synchronized audio and video.</p>
       </article>
       <article onclick="window.location.href='./image-decoder/animated-gif-renderer.html';">
