+++ samples/library/audio_renderer.js
@@ -1,4 +1,4 @@
-import { MP4PullDemuxer } from "./mp4_pull_demuxer.js";
+import { MP4PullDemuxer } from "../audio-video-player/mp4_pull_demuxer.js";
 import { RingBuffer } from "../third_party/ringbufjs/ringbuf.js";
 
 const DATA_BUFFER_DECODE_TARGET_DURATION = 0.3;
