+++ samples/library/audiosink.js
