+++ samples/samples.css
@@ -31,11 +31,11 @@ body {
   text-decoration: underline;
 }
 
-article a {
+a article {
   color: #FEFEF2;
   text-decoration: none;
 }
-article a:hover {
+a:hover article {
   color: white;
   text-decoration: underline;
 }
